review fixes II
