+++ mex/drop/main.py
@@ -1,20 +1,15 @@
 import reflex as rx
 
+from mex.drop.files_io import ALLOWED_CONTENT_TYPES
 from mex.drop.state import AppState, TempFile
-
-ACCEPTED_FILE_FORMATS = {
-    "application/json": [".json"],
-    "application/xml": [".xml"],
-    "text/csv": [".csv"],
-    "application/csv": [".csv"],
-    "text/tab-separated-values": [".tsv"],
-    "application/vnd.ms-excel": [".xls"],
-    "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": [".xlsx"],
-}
 
 
 def mex_drop_logo() -> rx.Component:
-    """Return the mex-drop logo with icon and label."""
+    """Return the mex-drop logo with icon and label.
+
+    Returns:
+        rx.Component: A component with style, event trigger and other props
+    """
     return rx.hstack(
         rx.icon(
             "droplets",
@@ -74,6 +69,96 @@ def create_file_row(temp_file: TempFile) -> rx.Component:
     )
 
 
+def create_drag_and_drop() -> rx.Component:
+    """Create card for drag and drop area for file selection."""
+    return rx.card(
+        rx.vstack(
+            mex_drop_logo(),
+            rx.divider(size="4"),
+            rx.text("Please select and upload your files here.", size="2"),
+            rx.upload(
+                rx.vstack(
+                    rx.icon(
+                        "file-down",
+                        size=30,
+                    ),
+                    rx.text(
+                        "Drag and drop " "or click to select files",
+                        size="1",
+                    ),
+                    rx.text(
+                        f"Supported formats: "
+                        f"{', '.join(ALLOWED_CONTENT_TYPES.values())}",
+                        size="1",
+                        color_scheme="gray",
+                    ),
+                    rx.button(
+                        "Select Files",
+                        color="var(--white-1)",
+                        bg="royalblue",
+                    ),
+                    align="center",
+                ),
+                id="upload_one",
+                max_files=100,
+                border="1px dotted var(--accent-8)",
+                padding="3em",
+                on_drop=AppState.handle_upload(rx.upload_files(upload_id="upload_one")),  # type: ignore[call-arg]
+            ),
+            spacing="4",
+        ),
+        top="20vh",
+        width="100%",
+        height="100%",
+        padding="15px",
+        custom_attrs={"data-testid": "index-card"},
+    )
+
+
+def create_file_handling_card() -> rx.Component:
+    """Create card for file handling and upload."""
+    return rx.card(
+        rx.vstack(
+            rx.scroll_area(
+                rx.flex(
+                    uploaded_file_display(),
+                ),
+                type="always",
+                scrollbars="vertical",
+                height=287,
+            ),
+            rx.form.root(
+                rx.hstack(
+                    rx.input(
+                        placeholder="API key",
+                        max_length=50,
+                        id="api_key",
+                        required=True,
+                    ),
+                    rx.input(
+                        placeholder="x-system",
+                        max_length=50,
+                        id="x-system",
+                        required=True,
+                    ),
+                    rx.button(
+                        "Submit",
+                        type="submit",
+                        width="20%",
+                        color="var(--white)",
+                        bg="royalblue",
+                    ),
+                ),
+                on_submit=AppState.submit_data,
+                reset_on_submit=False,
+            ),
+        ),
+        top="20vh",
+        width="100%",
+        height="100%",
+    )
+
+
 def index() -> rx.Component:
     """Return the index for the drop app.
 
@@ -84,99 +169,13 @@ def index() -> rx.Component:
         rx.center(
             rx.container(
                 rx.hstack(
-                    rx.card(
-                        rx.vstack(
-                            mex_drop_logo(),
-                            rx.divider(size="4"),
-                            rx.text(
-                                "Please select and upload your files here.", size="2"
-                            ),
-                            rx.upload(
-                                rx.vstack(
-                                    rx.icon(
-                                        "file-down",
-                                        size=30,
-                                    ),
-                                    rx.text(
-                                        "Drag and drop " "or click to select files",
-                                        size="1",
-                                    ),
-                                    rx.text(
-                                        "Supported formats: "
-                                        "JSON, XML, CSV, TSV, XLS and XLSX",
-                                        size="1",
-                                        color_scheme="gray",
-                                    ),
-                                    rx.button(
-                                        "Select Files",
-                                        color="white",
-                                        bg="royalblue",
-                                    ),
-                                    align="center",
-                                ),
-                                id="upload_one",
-                                max_files=100,
-                                accept=ACCEPTED_FILE_FORMATS,
-                                border="1px dotted rgb(107,99,246)",
-                                padding="3em",
-                                align="center",
-                                on_drop=AppState.handle_upload(
-                                    rx.upload_files(upload_id="upload_one")
-                                ),  # type: ignore[call-arg]
-                            ),
-                            spacing="4",
-                        ),
-                        top="20vh",
-                        width="100%",
-                        height="100%",
-                        padding="15px",
-                        custom_attrs={"data-testid": "index-card"},
-                    ),
-                    rx.card(
-                        rx.vstack(
-                            rx.scroll_area(
-                                rx.flex(
-                                    uploaded_file_display(),
-                                ),
-                                type="always",
-                                scrollbars="vertical",
-                                height=287,
-                            ),
-                            rx.form.root(
-                                rx.hstack(
-                                    rx.input(
-                                        placeholder="API key",
-                                        max_length=50,
-                                        id="api_key",
-                                        required=True,
-                                    ),
-                                    rx.input(
-                                        placeholder="x-system",
-                                        max_length=50,
-                                        id="x-system",
-                                        required=True,
-                                    ),
-                                    rx.button(
-                                        "Submit",
-                                        type="submit",
-                                        width="20%",
-                                        color="white",
-                                        bg="royalblue",
-                                    ),
-                                ),
-                                on_submit=AppState.submit_data,
-                                reset_on_submit=False,
-                            ),
-                        ),
-                        top="20vh",
-                        width="100%",
-                        height="100%",
-                    ),
+                    create_drag_and_drop(),
+                    create_file_handling_card(),
                     top="50vh",
                     width="100%",
                 )
             )
         ),
-        background_color="whitesmoke",
+        background_color="var(--gray-2)",
         min_height="100vh",
     )
